Move HasJsonContentType
